# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API for Xưởng Máy Gỗ Việt — woodworking machinery e-commerce
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export interface AdminUser {
  id: number;
  username: string;
  displayName: string;
}

export interface LoginResponse {
  token: string;
  admin: AdminUser;
}

export interface Category {
  slug: string;
  name: string;
  productCount: number;
}

export interface SpecItem {
  label: string;
  value: string;
}

export interface Product {
  id: number;
  slug: string;
  name: string;
  shortDescription: string;
  description: string;
  price: number;
  /** @nullable */
  salePrice?: number | null;
  currency: string;
  categorySlug: string;
  categoryName: string;
  brand: string;
  sku: string;
  /** @nullable */
  powerKw?: number | null;
  /** @nullable */
  weightKg?: number | null;
  /** @nullable */
  dimensions?: string | null;
  /** @nullable */
  warrantyMonths?: number | null;
  /** @nullable */
  origin?: string | null;
  coverImageUrl: string;
  gallery: string[];
  features: string[];
  specs: SpecItem[];
  inStock: boolean;
  featured: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProductRequest {
  slug: string;
  name: string;
  shortDescription: string;
  description: string;
  price: number;
  /** @nullable */
  salePrice?: number | null;
  currency: string;
  categorySlug: string;
  categoryName: string;
  brand: string;
  sku: string;
  /** @nullable */
  powerKw?: number | null;
  /** @nullable */
  weightKg?: number | null;
  /** @nullable */
  dimensions?: string | null;
  /** @nullable */
  warrantyMonths?: number | null;
  /** @nullable */
  origin?: string | null;
  coverImageUrl: string;
  gallery: string[];
  features: string[];
  specs: SpecItem[];
  inStock: boolean;
  featured: boolean;
}

export interface UpdateProductRequest {
  slug?: string;
  name?: string;
  shortDescription?: string;
  description?: string;
  price?: number;
  /** @nullable */
  salePrice?: number | null;
  currency?: string;
  categorySlug?: string;
  categoryName?: string;
  brand?: string;
  sku?: string;
  /** @nullable */
  powerKw?: number | null;
  /** @nullable */
  weightKg?: number | null;
  /** @nullable */
  dimensions?: string | null;
  /** @nullable */
  warrantyMonths?: number | null;
  /** @nullable */
  origin?: string | null;
  coverImageUrl?: string;
  gallery?: string[];
  features?: string[];
  specs?: SpecItem[];
  inStock?: boolean;
  featured?: boolean;
}

export interface Post {
  id: number;
  slug: string;
  title: string;
  excerpt: string;
  content: string;
  coverImageUrl: string;
  author: string;
  tags: string[];
  published: boolean;
  /** @nullable */
  publishedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreatePostRequest {
  slug: string;
  title: string;
  excerpt: string;
  content: string;
  coverImageUrl: string;
  author: string;
  tags: string[];
  published: boolean;
}

export interface UpdatePostRequest {
  slug?: string;
  title?: string;
  excerpt?: string;
  content?: string;
  coverImageUrl?: string;
  author?: string;
  tags?: string[];
  published?: boolean;
}

export interface HomeSummary {
  featuredProducts: Product[];
  latestPosts: Post[];
  categories: Category[];
  totalProducts: number;
}

export interface AdminStats {
  totalProducts: number;
  featuredProducts: number;
  outOfStockProducts: number;
  totalPosts: number;
  publishedPosts: number;
  totalCategories: number;
}

export type ListProductsParams = {
  category?: string;
  search?: string;
  featured?: boolean;
};

export type ListPostsParams = {
  search?: string;
};