# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API for Xưởng Máy Gỗ Việt — woodworking machinery e-commerce
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStats {
  totalProducts: number;
  featuredProducts: number;
  outOfStockProducts: number;
  totalPosts: number;
  publishedPosts: number;
  totalCategories: number;
}