# lib/api-zod/src/generated/types/createProductRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API for Xưởng Máy Gỗ Việt — woodworking machinery e-commerce
 * OpenAPI spec version: 0.1.0
 */
import type { SpecItem } from "./specItem";

export interface CreateProductRequest {
  slug: string;
  name: string;
  shortDescription: string;
  description: string;
  price: number;
  /** @nullable */
  salePrice?: number | null;
  currency: string;
  categorySlug: string;
  categoryName: string;
  brand: string;
  sku: string;
  /** @nullable */
  powerKw?: number | null;
  /** @nullable */
  weightKg?: number | null;
  /** @nullable */
  dimensions?: string | null;
  /** @nullable */
  warrantyMonths?: number | null;
  /** @nullable */
  origin?: string | null;
  coverImageUrl: string;
  gallery: string[];
  features: string[];
  specs: SpecItem[];
  inStock: boolean;
  featured: boolean;
}