# lib/api-zod/src/generated/types/loginResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API for Xưởng Máy Gỗ Việt — woodworking machinery e-commerce
 * OpenAPI spec version: 0.1.0
 */
import type { AdminUser } from "./adminUser";

export interface LoginResponse {
  token: string;
  admin: AdminUser;
}